Setting the date/time in the IoT Gateway test case
The IoT Gateway test case uses timestamped certificates
to connect to the AWS cloud. If the date/time on the dut
is set to before the certificates are valid then the test
case fails by saying 'the certificates are not yet valid'.

This commit addresses that issue by having the test case
set the time on the dut to the current date/time.

Signed-off-by: Jason Reeder <jreeder@ti.com>
Signed-off-by: Elvita Lobo <elvita@ti.com>