default_test_module.rb: Move translate_params2 to default module
translate_params2 is used to transform various assets for second
DUT to form a new boot params to pass to the booting login in
default_test_module. Since this transformation is not only to
pci tests and it could be used for any b-2-b testing, move this
part to default module. Also expand this function to transform
all the possible assets.

Signed-off-by: Yan Liu <yan-liu@ti.com>