sysfw: Improve log capture and handle resets
Some test reset the R5 requiring a reload of the firmware,
handle that use case via expected_resets/var_expected_resets
test parameter. Also some data was lost due to the time
it took between running the load script and re-connecting
to the R5, use a pipe instead of a load script so that
we have more control when to re-connect to the R5.

Signed-off-by: Alejandro Hernandez <ajhernandez@ti.com>