Fix default usb device speed to be high and add debug prints
USB device default speed was mentioned as "none" but this led to test cases
which do not mention speed in parameters to fail since speed has to be
low/high/superspeed. This patch fixes the default and it also adds some debug
prints to be logged into log files for easy debug.

Signed-off-by: Aparna Balasubramanian <aparnab@ti.com>
Signed-off-by: Yan Liu <yan-liu@ti.com>