LSP: A-Display: drm: Refactor get_test_modes function
The function used to generate test modes created an
exhaustive list of combinations for the multi-display
modes with newer hardware that support 2k, 4k and
higher resolution modes the list of modes can grow to
more than 1000 test modes. With 1k+ modes to validate
the increase in test time is too high so refactor the
function so that fewer test modes are generated by the
function but ensure that the list of test modes created
includes each mode supported by each connector.

Signed-off-by: Alejandro Hernandez <ajhernandez@ti.com>