scmcsdk: tidl_import_tool_regression: Fix bug calling demo
The script does a git clone on the server side that takes few minutes
and when it goes back and tries to execute the demo on the DUT, the
session has already timed out and user is at login prompt.
Run the git clone in the target instead of the server to avoid
session timeout.

Signed-off-by: Carlos Hernandez <ceh@ti.com>